% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\docType{data}
\name{example_data}
\alias{example_data}
\title{Sample small dataset to test the package}
\format{An object of class \code{list} of length 3.}
\usage{
example_data
}
\description{
Contains 'n' observations of 's' continuous responses
and 'p' SNPs-like covariates
}
\details{
n = 100 ; p = 150 ; s = 10

Loading the data will load the associated 
blockList and strucutreGraph
needed to run the algorithm.
}
\keyword{datasets}
