% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data_ground_thruth.R
\docType{data}
\name{example_ground_truth}
\alias{example_ground_truth}
\title{Ground truth to compare results of estimation in the
sample dataset provided in example_data.rda}
\format{An object of class \code{list} of length 2.}
\usage{
example_ground_truth
}
\description{
Contains both the true gamma (variable selection for covariates)
and the graph G (covariance selection for the residuals)
}
\keyword{datasets}
