\name{R2SSUR-package}
\alias{R2SSUR-package}
\alias{R2SSUR}
\docType{package}
\title{\packageTitle{R2SSUR}}
\description{\packageDescription{R2SSUR}}
\details{
  The DESCRIPTION file: \packageDESCRIPTION{R2SSUR}
  \packageIndices{R2SSUR}

  This section should provide a more detailed overview of how to use the
  package, including the most important functions.
}
\author{
  \packageAuthor{R2SSUR}
  
  Maintainer: \packageMaintainer{R2SSUR}
}
\references{
  This optional section can contain literature or other references for
  background information.
}
% Optionally other standard keywords, one per line,
% from the file KEYWORDS in the R documentation.
\keyword{package}
\seealso{
  Optional links to other man pages
}
\examples{
  ## Optional simple examples of the most important functions
  ## Use \dontrun{} around code to be shown but not executed
}
