% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runSSUR.R
\name{runSSUR}
\alias{runSSUR}
\alias{my_stop}
\title{runSSUR}
\usage{
my_stop(msg, tmpFolder)
}
\arguments{
\item{tmpFolder}{the path to a temporary folder where intermediate data files are stored (will be erased at the end of the chain) default to local tmpFolder}

\item{data}{either a matrix/dataframe or the path to (a plain text) data file with variables on the columns and observations on the rows}

\item{Y, X, X_0}{vectors of indexes (with respect to the data matrix) for the outcomes, the covariates to select and the fixed covariates respectively if data is either a path to a file or a matrix;
if the 'data' argument is not provided, these needs to be matrices containing the data instead.}

\item{varType}{variable type for each column in the data file coded as: 0 - continuous, 1- binary, 2 - categorical. Note that categorical variables cannot be imputed}

\item{outFilePath}{path to where the output files are to be written}

\item{nIter}{number of iterations for the MCMC procedure}

\item{burnin}{number of iterations (or fraction of iterations) to discard at the start of the chain Default = 0}

\item{nChains}{number of parallel chains to run}

\item{covariancePrior}{string indicating the prior for the covariance $C$; it has to be either "HIW" for the hyper-inverse-Wishar (which will result in a sparse covariance matrix),
"IW" for the inverse-Wishart prior ( dense covariance ) or "IG" for independent inverse-Gamma on all the diagonal elements and 0 otherwise.}

\item{gammaPrior}{string indicating the gamma prior to use, either "hotspot" for the Hotspot prior of Bottolo (2011), "MRF" for the Markov Random Field prior or "hierarchical" for a simpler hierarchical prior}

\item{gammaSampler}{string indicating the type of sampler for gamma, either "bandit" for the Thompson sampling inspired samper or "MC3" for the usual $MC^3$ sampler}

\item{gammaInit}{gamma initialisation to either all-zeros ("0"), all ones ("1"), randomly ("R") or (default) MLE-informed ("MLE").}

\item{mrfG}{either a matrix or a path to the file containing the G matrix for the MRF prior on gamma (if necessary)}

\item{hyperpar}{a list of named hypeparameters to use instead of the default values}
}
\description{
Run a SUR Bayesian sampler
}
\details{
R2SSUR -- Bayesian Sparse Seemingly Unrelated Regression
}
\examples{
\donttest{
data(example_data, package = "R2SSUR")

# set some non-default hyperparameters
hyperpar = list( a_w = 2 , b_w = 5 )

fit = R2SSUR::runSSUR(example_data[["data"]],outFilePath = "results/",
                      Y = example_data[["blockList"]][[1]],
                      X = example_data[["blockList"]][[2]],
                      nIter = 100, nChains = 2, method = "SUR", gammaPrior = "hotspot",hyperpar = hyperpar )

## check output
greyscale = grey((100:0)/100)
data(example_ground_truth, package = "R2SSUR")

est_gamma = as.matrix( read.table(fit$output$gamma) )
est_G = as.matrix( read.table(fit$output$G) )
s = ncol(est_G)

par(mfrow=c(2,2))
image(est_gamma,col=greyscale) image(example_ground_truth[["gamma"]],col=greyscale)
image((est_G+diag(s))[s:1,],col=greyscale) image(example_ground_truth[["G"]][s:1,],col=greyscale)
par(mfrow=c(1,1))
}

}
