% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runSSUR.R
\name{runSSUR}
\alias{runSSUR}
\title{runSSUR}
\usage{
runSSUR(data, blockList, varType = NULL, structureGraph,
  outFilePath = "", nIter = 10, nChains = 1, method = "SSUR",
  gammaSampler = "Bandit", gammaInit = "MLE", usingGPrior = FALSE)
}
\arguments{
\item{data}{path to data file}

\item{outFilePath}{path to where the output is to be written}

\item{nIter}{number of iterations}

\item{nChains}{number of parallel chains to run}
}
\description{
Run a SUR Bayesian sampler
}
\details{
R2SSUR -- Bayesian Sparse Seemingly Unrelated Regression
}
\examples{
require(utils)
dir.create("tmp")
# ** code to call the function on sample data here **
unlink("tmp", recursive=TRUE)

}
