% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runSSUR.R
\name{runSSUR}
\alias{runSSUR}
\title{runSSUR}
\usage{
runSSUR(data, blockList, varType = NULL, structureGraph = NULL,
  outFilePath = "", nIter = 10, burnin = 0, nChains = 1,
  method = "SSUR", gammaSampler = "Bandit", gammaInit = "MLE",
  usingGPrior = FALSE)
}
\arguments{
\item{data}{either a matrix/dataframe or the path to (a plain text) data file with variables on the columns and observations on the rows}

\item{blockList}{list of blocks in the model; each element of the list contains the (column) indices of the variables in each block, with respect to the data file}

\item{varType}{variable type for each column in the data file; coded as: 0 - continuous, 1- binary, 2 - categorical. Note that categorical variables cannot be imputed}

\item{structureGraph}{graph adjacency matrix representing the structure between the blocks. Edges represented as 2 indicate variables that will be always included in the regression model}

\item{outFilePath}{path to where the output files are to be written}

\item{nIter}{number of iterations for the MCMC procedure}

\item{burnin}{number of iterations (or fraction of iterations) to discard at the start of the chain; Default = 0}

\item{nChains}{number of parallel chains to run}

\item{method}{a string indicating the model type, either "SSUR" or "dSUR" for sparse and dense Seemingly Unrelated Regressions, or "HESS" for independent outcomes.}

\item{gammaInit}{gamma initialisation to either all-zeros ("0"), all ones ("1"), randomly ("R") or (default) MLE-informed ("MLE").}
}
\description{
Run a SUR Bayesian sampler
}
\details{
R2SSUR -- Bayesian Sparse Seemingly Unrelated Regression
}
\examples{
\donttest{
data(example_data, package = "R2SSUR")

R2SSUR::runSSUR(example_data[["data"]],outFilePath = "results/",
                blockList = example_data[["blockList"]], structureGraph = example_data[["structureGraph"]],
                nIter = 20000,nChains = 2, method = "SSUR")

## check output
greyscale = grey((1000:0)/1000)
data(example_ground_truth, package = "R2SSUR")

est_gamma = as.matrix( read.table("results/data_SSUR_gamma_out.txt") )
est_G = as.matrix( read.table("results/data_SSUR_G_out.txt") )
s = ncol(est_G)

par(mfrow=c(2,2))
image(est_gamma,col=greyscale); image(example_ground_truth[["gamma"]],col=greyscale)
image((est_G+diag(s))[s:1,],col=greyscale); image(example_ground_truth[["G"]][s:1,],col=greyscale)
par(mfrow=c(1,1))
}

}
