% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runSSUR.R
\name{runSSUR}
\alias{runSSUR}
\title{runSSUR}
\usage{
runSSUR(data, blockList, varType = NULL, structureGraph = NULL,
  outFilePath = "", nIter = 10, nChains = 1, method = "SSUR",
  gammaSampler = "Bandit", gammaInit = "MLE", usingGPrior = FALSE)
}
\arguments{
\item{data}{path to data file}

\item{outFilePath}{path to where the output is to be written}

\item{nIter}{number of iterations}

\item{nChains}{number of parallel chains to run}
}
\description{
Run a SUR Bayesian sampler
}
\details{
R2SSUR -- Bayesian Sparse Seemingly Unrelated Regression
}
\examples{
\donttest{
data(example_data, package = "R2SSUR")

R2SSUR::runSSUR(example_data[["data"]],outFilePath = "results/",
                blockList = example_data[["blockList"]], structureGraph = example_data[["structureGraph"]],
                nIter = 20000,nChains = 2, method = "SSUR")

## check output
greyscale = grey((1000:0)/1000)
data(example_ground_truth, package = "R2SSUR")

est_gamma = as.matrix( read.table("results/data_SSUR_gamma_out.txt") )
est_G = as.matrix( read.table("results/data_SSUR_G_out.txt") )
s = ncol(est_G)

par(mfrow=c(2,2))
image(est_gamma,col=greyscale); image(example_ground_truth[["gamma"]],col=greyscale)
image((est_G+diag(s))[s:1,],col=greyscale); image(example_ground_truth[["G"]][s:1,],col=greyscale)
par(mfrow=c(1,1))
}

}
